+++ src/layers/ImageLayer.js
@@ -10,7 +10,7 @@ const defaultProps = {
   pickable: { type: 'boolean', value: true, compare: true },
   coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
   contrastLimits: { type: 'array', value: [], compare: true },
-  active: { type: 'array', value: [], compare: true },
+  channelsVisible: { type: 'array', value: [], compare: true },
   colors: { type: 'array', value: [], compare: true },
   selections: { type: 'array', value: [], compare: true },
   colormap: { type: 'string', value: '', compare: true },
@@ -44,7 +44,7 @@ const defaultProps = {
  * @type {Object}
  * @property {Array.<Array.<number>>} contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @property {Array.<Array.<number>>} colors List of [r, g, b] values for each channel.
- * @property {Array.<boolean>} active List of boolean values for each channel for whether or not it is visible.
+ * @property {Array.<boolean>} channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @property {Object} loader PixelSource. Represents an N-dimensional image.
  * @property {Array} selections Selection to be used for fetching data.
  * @property {number=} opacity Opacity of the layer.