+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js
@@ -16,7 +16,7 @@ const defaultProps = {
   onHover: { type: 'function', value: null, compare: false },
   contrastLimits: { type: 'array', value: [], compare: true },
   colors: { type: 'array', value: [], compare: true },
-  active: { type: 'array', value: [], compare: true },
+  channelsVisible: { type: 'array', value: [], compare: true },
   opacity: { type: 'number', value: 1, compare: true },
   colormap: { type: 'string', value: '', compare: true },
   domain: { type: 'array', value: [], compare: true },
@@ -38,7 +38,7 @@ const defaultProps = {
  * @type {object}
  * @property {Array.<Array.<number>>} contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @property {Array.<Array.<number>>} colors List of [r, g, b] values for each channel.
- * @property {Array.<boolean>} active List of boolean values for each channel for whether or not it is visible.
+ * @property {Array.<boolean>} channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @property {Array} loader Image pyramid. PixelSource[], where each PixelSource is decreasing in shape.
  * @property {Array} selections Selection to be used for fetching data.
  * @property {number=} opacity Opacity of the layer.