+++ src/layers/OverviewLayer.js
@@ -40,7 +40,7 @@ const defaultProps = {
  * @type {Object}
  * @property {Array.<Array.<number>>} contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @property {Array.<Array.<number>>} colors List of [r, g, b] values for each channel.
- * @property {Array.<boolean>} active List of boolean values for each channel for whether or not it is visible.
+ * @property {Array.<boolean>} channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @property {Array} loader PixelSource[]. Assumes multiscale if loader.length > 1.
  * @property {Array} selections Selection to be used for fetching data.
  * @property {number=} opacity Opacity of the layer.