+++ src/layers/VolumeLayer/VolumeLayer.js
@@ -11,7 +11,7 @@ const defaultProps = {
   pickable: false,
   coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
   contrastLimits: { type: 'array', value: [], compare: true },
-  active: { type: 'array', value: [], compare: true },
+  channelsVisible: { type: 'array', value: [], compare: true },
   colors: { type: 'array', value: [], compare: true },
   colormap: { type: 'string', value: '', compare: true },
   selections: { type: 'array', value: [], compare: true },
@@ -48,7 +48,7 @@ const defaultProps = {
  * @type {Object}
  * @property {Array.<Array.<number>>} contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @property {Array.<Array.<number>>} colors List of [r, g, b] values for each channel.
- * @property {Array.<boolean>} active List of boolean values for each channel for whether or not it is visible.
+ * @property {Array.<boolean>} channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @property {Array} loader PixelSource[]. Represents an N-dimensional image.
  * @property {Array} selections Selection to be used for fetching data.
  * @property {string=} colormap String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap