+++ src/layers/XR3DLayer/XR3DLayer.js
@@ -117,7 +117,7 @@ function removeExtraColormapFunctionsFromShader(colormap) {
  * @type {Object}
  * @property {Array.<Array.<number>>} contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @property {Array.<Array.<number>>} colors List of [r, g, b] values for each channel.
- * @property {Array.<boolean>} active List of boolean values for each channel for whether or not it is visible.
+ * @property {Array.<boolean>} channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @property {string} dtype Dtype for the layer.
  * @property {string=} colormap String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
  * @property {Array.<Array.<number>>=} domain Override for the possible max/min values (i.e something different than 65535 for uint16/'<u2').
@@ -240,7 +240,7 @@ const XR3DLayer = class extends Layer {
       ySlice,
       zSlice,
       modelMatrix,
-      active,
+      channelsVisible,
       domain,
       dtype,
       clippingPlanes,
@@ -255,7 +255,7 @@ const XR3DLayer = class extends Layer {
       const { paddedContrastLimits, paddedColors } = padColorsAndWindows({
         contrastLimits,
         colors,
-        active,
+        channelsVisible,
         domain,
         dtype
       });