+++ src/viewers/PictureInPictureViewer.jsx
@@ -15,7 +15,7 @@ import { GLOBAL_SLIDER_DIMENSION_FIELDS } from '../constants';
  * @param {Object} props
  * @param {Array} props.contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @param {Array} props.colors List of [r, g, b] values for each channel.
- * @param {Array} props.active List of boolean values for each channel for whether or not it is visible.
+ * @param {Array} props.channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @param {string} [props.colormap] String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
  * @param {Array} props.loader The data source for the viewer, PixelSource[]. If loader.length > 1, data is assumed to be multiscale.
  * @param {Array} props.selections Selection to be used for fetching data.
@@ -51,7 +51,7 @@ const PictureInPictureViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     viewStates: viewStatesProp,
     colormap,
     overview,
@@ -96,7 +96,7 @@ const PictureInPictureViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     selections: oldselections,
     newselections,
     onViewportLoad,