+++ src/viewers/SideBySideViewer.jsx
@@ -9,7 +9,7 @@ import { GLOBAL_SLIDER_DIMENSION_FIELDS } from '../constants';
  * @param {Object} props
  * @param {Array} props.contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @param {Array} props.colors List of [r, g, b] values for each channel.
- * @param {Array} props.active List of boolean values for each channel for whether or not it is visible.
+ * @param {Array} props.channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @param {string} [props.colormap] String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
  * @param {Array} props.loader This data source for the viewer. PixelSource[]. If loader.length > 1, data is assumed to be multiscale.
  * @param {Array} props.selections Selection to be used for fetching data.
@@ -39,7 +39,7 @@ const SideBySideViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     viewStates: viewStatesProp,
     colormap,
     panLock,
@@ -103,7 +103,7 @@ const SideBySideViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     selections: oldselections,
     newselections,
     onViewportLoad,