+++ src/viewers/VolumeViewer.jsx
@@ -9,7 +9,7 @@ import { RENDERING_MODES } from '../constants';
  * @param {Object} props
  * @param {Array} props.contrastLimits List of [begin, end] values to control each channel's ramp function.
  * @param {Array} props.colors List of [r, g, b] values for each channel.
- * @param {Array} props.active List of boolean values for each channel for whether or not it is visible.
+ * @param {Array} props.channelsVisible List of boolean values for each channel for whether or not it is visible.
  * @param {string} [props.colormap] String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
  * @param {Array} props.loader This data source for the viewer. PixelSource[]. If loader.length > 1, data is assumed to be multiscale.
  * @param {Array} props.selections Selection to be used for fetching data
@@ -34,7 +34,7 @@ const VolumeViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     selections,
     colormap,
     resolution = Math.max(0, loader.length - 1),
@@ -80,7 +80,7 @@ const VolumeViewer = props => {
     loader,
     contrastLimits,
     colors,
-    active,
+    channelsVisible,
     selections,
     colormap,
     xSlice,