+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 - Change API for all public props
   - `sliderValues` -> `contrastLimits`
   - `colorValues` -> `colors`
-  - `channelIsOn` -> `active`
+  - `channelIsOn` -> `channelsVisible`
   - `loaderSelection` -> `selections`
   - `MAX_SLIDERS_AND_CHANNELS` -> `MAX_CHANNELS`